refactor(test): extract click_canvas_fraction and submit_label_dialog
